refactor: Gemini client to implement BaseLLMClient interface
Move files to providers/gemini structure and update imports
Update default model name to gemini-2.0-flash
Implement generate method for LLMRequest/Response compatibility
Add proper error handling with API key redaction
